Replace mentors.js and mentors.json with mentors_copy.js and mentors_new.json
